MANY MANY CHANGES I. Regarding Android Client  1. Tried to implement onSaveInstance(). However, this was tricky because I was trying to save my messages arraylist,     which happens to be an ArrayList<MessageItem>. HOWEVER, because MessageItem is not a primitive type, things got tricky.     I had to make MessageItem implement Parcelable, and add a bunch of methods.     It still isn't working.
 2. ClientOutgoingTask():
    Still working on generating a date, but haven't come close yet.

 3. CustomListViewAdapter():
    Implemented date formatting. Depending on how far the timestamp is from the current time, the date will show up
    differently on the screen. Ran into a HUGE bug <see below>

Basically, it kept returning (Month) (Day) whenever I was testing. And when I mean testing, I mean that
I had the server running on the pi, and a c++ client running on cygwin, which would send the messages. The Android client would receive them.
The Android client was receiving the messages just fine, but I kept getting (Month) (Day). So then I was like, ok that's weird, that means that
the difference in time between System.currentTimeMillis() and the timestamp is greater than 1. So I printed the raw difference.
And turns out, it was a hugeass number! I was like, whaaaat. So I printed out both System.currentTimeMillis() and the timestamp.
It turns out, the timestamp was WAAAY WAAAY behind. So I was like, okay, it's a problem in the c++ client send, particularly ClientSender's message generator.
So I looked at that, and it turns out, what I was using (std::time(NULL)) only returned time from the epoch in SECONDS, not milliseconds. Now, I'm not just going to
naively multiply that by 1000. Too much loss of precision.

So I spent literally 2 days looking for a viable solution. Until finally, finally, I found something.
http://en.cppreference.com/w/cpp/chrono/time_point/time_since_epoch

Basically, it uses a bunch of methods from std::chrono.

..with a few modifications. Instead of using std::chrono::hours, I used std::chrono::milliseconds.
Also, I found out that in order to get the correct value, it would need to be stored as a long long, not a long.
So I changed THE ENTIRE PROJECT IMPLEMENTATION to take long long, instead of long. THis meant making modifying:
 1. ClientSender.cpp, h
 2. ClientReceiver.cpp, h
 3. ClientTimestampFiller.cpp, h
 3. MessageItem
 4. ConnectionHandler
 5. MessageHandler

So I was like good, now everything should be working right. Nope. The build wasn't even compiling.
It was then that I realized that std::chrono was c++11, and my makefiles didn't specify any specific flags.
So I went and added in -std=c++11 -std=gnu++11 to CFLAGS of both my Makefiles

+++ mtserver/ClientReceiver.cpp
@@ -1,7 +1,7 @@
 #include "ClientReceiver.h"
 #define MAX_MESSAGE_SIZE 25600
 
-ClientReceiver::ClientReceiver(TCPStream* s, wqueue<long>& queue) : stream(s), m_queue(queue) {}
+ClientReceiver::ClientReceiver(TCPStream* s, wqueue<long long>& queue) : stream(s), m_queue(queue) {}
 
 /*
 	Receives incoming messages from the server
@@ -39,7 +39,7 @@ void* ClientReceiver::run() {
 			std::string delimiter = ":::::::";
 
 			//Fields to extract
-			long timestamp;
+			long long timestamp;
 			std::string date_formatted;
 			std::string message;
 			std::string sender;
@@ -50,7 +50,7 @@ void* ClientReceiver::run() {
 				std::cout << "Regular message, Item found: " << current_item << std::endl;
 				switch (which_one) {
 				case 0:		//timestamp
-					timestamp = atol(current_item.c_str());
+					timestamp = atoll(current_item.c_str());
 					break;
 				case 1:		//formatted date
 					date_formatted = current_item;
